requirements.txt
# ── Worker deps (required for scanner_worker + quantflow_scanner_core) ─────────
pandas>=2.0
numpy>=1.24
requests>=2.31
PyYAML>=6.0

# ── Streamlit app deps (required for app.py) ────────────────────────────────────
streamlit>=1.35
plotly>=5.20
